исправление адресов rest api в projects-service, перемещение пакетов в пакет projects, создание файла Constants
