--- students-registry/frontend/src/main/java/boost/brain/course/controller/MainController.java
@@ -1,15 +1,0 @@
-package boost.brain.course.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/")
-public class MainController {
-
-    @GetMapping
-    public String main(){
-        return "index.html";
-    }
-}
