+++ students-registry/frontend/src/main/java/boost/brain/course/projects/controller/MainController.java
@@ -1,0 +1,15 @@
+package boost.brain.course.projects.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/")
+public class MainController {
+
+    @GetMapping
+    public String main(){
+        return "index.html";
+    }
+}
