--- students-registry/projects-service/src/main/java/boost/brain/course/configuration/SwaggerConfig.java
@@ -1,28 +1,0 @@
-package boost.brain.course.configuration;
-
-import boost.brain.course.model.ProjectMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
-
-@Configuration
-@EnableSwagger2
-@ComponentScan(basePackageClasses = ProjectMapper.class)
-public class SwaggerConfig {
-    @Bean
-    public Docket api() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.any())
-                .paths(PathSelectors.any())
-                .build();
-    }
-
-  //  @Bean
-  //  public ProjectMapper projectMapper(){return new ProjectMapper();}
-}
