+++ students-registry/projects-service/src/main/java/boost/brain/course/projects/Constants.java
@@ -1,0 +1,18 @@
+package boost.brain.course.projects;
+
+public interface Constants {
+    String API_PREFIX = "/api";
+
+    String PROJECTS_CONTROLLER_PREFIX = API_PREFIX + "/projects";
+
+    String CREATE_PREFIX = "/create";
+    String READ_PREFIX = "/read";
+    String UPDATE_PREFIX = "/update";
+    String DELETE_PREFIX = "/delete";
+
+    String COUNT_PREFIX = "/count";
+    String FOR_IDS_PREFIX = "/for-ids";
+    String ALL = "/all";
+    String IF_EXISTS = "/if-exists";
+    String PAGE = "/page";
+}
