+++ students-registry/projects-service/src/main/java/boost/brain/course/projects/controller/exceptions/BadRequestException.java
@@ -1,0 +1,8 @@
+package boost.brain.course.projects.controller.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
+public class BadRequestException extends RuntimeException{
+}
