+++ students-registry/projects-service/src/main/java/boost/brain/course/projects/controller/exceptions/NotFoundException.java
@@ -1,0 +1,8 @@
+package boost.brain.course.projects.controller.exceptions;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class NotFoundException extends RuntimeException {
+}
