--- students-registry/users-service-test-util/src/main/java/boost/brain/course/controller/InputOperations.java
@@ -1,15 +1,0 @@
-package boost.brain.course.controller;
-
-public enum InputOperations {
-    COUNT,
-    CREATE,
-    DELETE,
-    PAGE,
-    READ,
-    UPDATE,
-    USERS_FOR_EMAILS,
-    USERS_ALL,
-    CHECK_IF_EXISTS,
-    UPDATE_STATUS,
-    UPDATE_STATUSES_FOR_EMAILS
-}
