+++ students-registry/users-service-test-util/src/main/java/boost/brain/course/projects/controller/InputOperations.java
@@ -1,0 +1,15 @@
+package boost.brain.course.projects.controller;
+
+public enum InputOperations {
+    COUNT,
+    CREATE,
+    DELETE,
+    PAGE,
+    READ,
+    UPDATE,
+    USERS_FOR_EMAILS,
+    USERS_ALL,
+    CHECK_IF_EXISTS,
+    UPDATE_STATUS,
+    UPDATE_STATUSES_FOR_EMAILS
+}
